change name from MemoryTable to MemorySource
