+++ server/src/main/java/org/apache/iotdb/db/mpp/plan/execution/memory/StatementMemorySourceVisitor.java
@@ -20,22 +20,17 @@
 package org.apache.iotdb.db.mpp.plan.execution.memory;
 
 import org.apache.iotdb.db.mpp.common.header.DatasetHeader;
+import org.apache.iotdb.db.mpp.plan.statement.StatementNode;
+import org.apache.iotdb.db.mpp.plan.statement.StatementVisitor;
 import org.apache.iotdb.tsfile.read.common.block.TsBlock;
 
-public class StatementMemoryTable {
-  private final TsBlock tsBlock;
-  private final DatasetHeader datasetHeader;
+import java.util.ArrayList;
 
-  public StatementMemoryTable(TsBlock tsBlock, DatasetHeader datasetHeader) {
-    this.tsBlock = tsBlock;
-    this.datasetHeader = datasetHeader;
-  }
+public class StatementMemorySourceVisitor
+    extends StatementVisitor<StatementMemorySource, StatementMemorySourceContext> {
 
-  public TsBlock getTsBlock() {
-    return tsBlock;
-  }
-
-  public DatasetHeader getDatasetHeader() {
-    return datasetHeader;
+  @Override
+  public StatementMemorySource visitNode(StatementNode node, StatementMemorySourceContext context) {
+    return new StatementMemorySource(new TsBlock(0), new DatasetHeader(new ArrayList<>(), false));
   }
 }
